+++ packages/cmf-router/src/sagaRouter.js
@@ -157,7 +157,7 @@ export default function* sagaRouter(history, routes) {
 	const routeFragments = Object.keys(routes);
 	while (true) {
 		const shouldStart = [];
-		const currentLocation = history.getCurrentLocation();
+		const currentLocation = history.location;
 		for (let index = 0; index < routeFragments.length; ) {
 			const routeFragment = routeFragments[index];
 			const routeSaga = routes[routeFragment];