+++ packages/datagrid/src/components/PinHeaderRenderer/PinHeaderRenderer.component.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { ButtonIcon } from '@talend/design-system';
-import theme from './PinHeaderRenderer.scss';
 
-export const PIN_HEADER_RENDERER_COMPONENT = 'pinHeaderRenderer';
+import theme from './PinHeaderRenderer.scss';
 
 export type PinHeaderRendererProps = Omit<Parameters<typeof ButtonIcon>[0], 'size' | 'icon'>;