+++ tools/scripts-config-react-webpack/config/webpack.config.js
@@ -319,7 +319,7 @@ module.exports = ({ getUserConfig, mode }) => {
 				publicPath: '/',
 				globalObject: 'this',
 			},
-			devtool: 'cheap-module-source-map',
+			devtool: 'source-map',
 			resolve: {
 				extensions: ['.js', useTypescript && '.ts', useTypescript && '.tsx'].filter(Boolean),
 			},