+++ packages/cmf-cqrs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @talend/react-cmf-cqrs
 
+## 7.2.0
+
+### Minor Changes
+
+- 47b758112: feat(ARCH-482): use React 17 internally and extend react peer dep version
+
+### Patch Changes
+
+- Updated dependencies [47b758112]
+  - @talend/react-cmf@7.1.0
+
 ## 7.1.1
 
 ### Patch Changes