+++ packages/components/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @talend/react-components
 
+## 6.49.0
+
+### Minor Changes
+
+- 47b758112: feat(ARCH-482): use React 17 internally and extend react peer dep version
+
+### Patch Changes
+
+- 6562e397f: fix(components): datalist selection by id should allow all types
+- Updated dependencies [47b758112]
+  - @talend/react-bootstrap@1.35.0
+  - @talend/design-system@2.7.0
+  - @talend/icons@6.42.0
+  - @talend/bootstrap-theme@6.39.0
+  - @talend/utils@2.3.0
+
+## 6.48.0
+
+### Minor Changes
+
+- bdf23018f: Add a tooltip to the tag
+
+## 6.47.0
+
+### Minor Changes
+
+- 7d1f9f509: Add a Feature tag on drawer subtitle
+- 66fa02548: move from react-bootstrap to @talend/react-bootstrap
+
+### Patch Changes
+
+- Updated dependencies [66fa02548]
+  - @talend/react-bootstrap@1.34.0
+
+## 6.46.7
+
+### Patch Changes
+
+- ecd3b81ae: fix(components): datalist should show selected item by id
+
 ## 6.46.6
 
 ### Patch Changes