+++ packages/faceted-search/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 4.2.0
+
+### Minor Changes
+
+- 47b758112: feat(ARCH-482): use React 17 internally and extend react peer dep version
+
+### Patch Changes
+
+- Updated dependencies [47b758112]
+  - @talend/react-bootstrap@1.35.0
+
+## 4.1.0
+
+### Minor Changes
+
+- 66fa02548: move from react-bootstrap to @talend/react-bootstrap
+
+### Patch Changes
+
+- Updated dependencies [66fa02548]
+  - @talend/react-bootstrap@1.34.0
+
 ## 4.0.2
 
 ### Patch Changes