+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @talend/icons
 
+## 6.42.0
+
+### Minor Changes
+
+- 47b758112: feat(ARCH-482): use React 17 internally and extend react peer dep version
+
+## 6.41.1
+
+### Patch Changes
+
+- 293a4ebd9: feat(icons): expose icons from figma as extra svg sprites
+
 ## 6.41.0
 
 ### Minor Changes